workers/banner/pyproject.toml
[project]
name = "dqx-banner-worker"
version = "1.0.0"
description = "DQX Banner translation API - Translate Dragon Quest X hiroba banners"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-vision>=3.10.1",
    "httpx>=0.28.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "openai>=1.66.5",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pydantic>=2.0.0",
    "scikit-learn>=1.6.1",
    "shapely>=2.0.7",
    "textdistance>=4.6.3",
]

[dependency-groups]
dev = [
    "workers-py",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
